# api-gateway/docs/swagger.yaml
definitions:
...
  user_proto.ListUser:
    properties:
      user:
        items:
          $ref: '#/definitions/user_proto.User'
        type: array
    type: object
  user_proto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  user_proto.LoginResponse:
    properties:
      token:
        type: string
    type: object
  user_proto.UpdatePasswordReq:
    properties:
      id:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    type: object
  user_proto.UpdateUserReq:
    properties:
      email:
        type: string
      id:
        type: string
      price:
        type: number
      username:
        type: string
    type: object
  user_proto.UpdateUserRes:
    properties:
      message:
        type: string
    type: object
  user_proto.User:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
      price:
        type: number
      username:
        type: string
    type: object
  user_proto.UserRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  user_proto.UserResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
  user_proto.VerifyReq:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server for a restaurant reservation system.
  title: Flash-Sale
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Log in a user with email and password
      parameters:
      - description: Login request body
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/user_proto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_proto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: User login
      tags:
      - user
...
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user_proto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_proto.UpdateUserRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new user
      tags:
      - user
  /users:
    get:
      description: Retrieve all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_proto.ListUser'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get all users
      tags:
      - user
  /users/{id}:
    delete:
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_proto.UpdateUserRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete a user
      tags:
      - user
    get:
      description: Get a user by their unique ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_proto.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update a user's details by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update request body
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/user_proto.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_proto.UpdateUserRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update a user
      tags:
      - user
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Update a user's password by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Password update request body
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/user_proto.UpdatePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_proto.UpdateUserRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update user password
      tags:
      - user
  /verify-code:
    post:
      consumes:
      - application/json
      description: Verify the user code sent to the user's email
      parameters:
      - description: Verification request body
        in: body
        name: verify
        required: true
        schema:
          $ref: '#/definitions/user_proto.VerifyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_proto.UserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Verify a user code
      tags:
      - user
securityDefinitions:
  Bearer:
    description: Enter the token in the format `Bearer {token}`
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"